all sql implementation, i hope everything works part 2(we refactored the vote route and function)

+++ server.py
@@ -4,7 +4,6 @@ import data_manager
 import operator
 import os
 from operator import itemgetter
-import time
 
 
 app = Flask(__name__)
@@ -14,8 +13,8 @@ photos = UploadSet('photos', IMAGES)
 app.config['UPLOADED_PHOTOS_DEST'] = "static/images"
 configure_uploads(app, photos)
 
-#q_head = data_manager.QUESTION_HEADERS
-#a_head = data_manager.ANSWER_HEADERS
+q_head = data_manager.QUESTION_HEADERS
+a_head = data_manager.ANSWER_HEADERS
 
 
 @app.route('/')
@@ -33,50 +32,58 @@ def index():
 @app.route('/add-question', methods=['POST'])
 def add_question():
     submission_time = data_manager.time_generator()
-    return render_template("question.html", next_id=0, submission_time=submission_time)
+
+    return render_template("question.html", submission_time=submission_time)
 
 
 @app.route("/question/<question_id>/delete")
 def delete_question(question_id):
     data = request.args.to_dict()
-    os.remove(os.path.join(app.config['UPLOADED_PHOTOS_DEST'], data['image']))
+    if data['image'] != '':
+        os.remove(os.path.join(app.config['UPLOADED_PHOTOS_DEST'], data['image']))
     data_manager.delete(question_id)
     return redirect('/list')
 
 
-@app.route("/answer/<answer_id>/delete")
+@app.route("/answer/<int:answer_id>/delete")
 def delete_answer(answer_id):
     question_id = data_manager.get_question_id_by_answer_id(answer_id)
     data = request.args.to_dict()
-    os.remove(os.path.join(app.config['UPLOADED_PHOTOS_DEST'], data['image']))
+    if data['image'] != '':
+        os.remove(os.path.join(app.config['UPLOADED_PHOTOS_DEST'], data['image']))
     data_manager.delete_answer(answer_id)
-    return redirect('question/{}'.format(question_id))
+    return redirect('question/{}'.format(question_id['question_id']))
+
+
+@app.route("/answer/", methods=['POST'])
+@app.route("/question/", methods=['POST'])
+def save_question():
+    if request.method == 'POST' and 'photo' in request.files:
+        try:
+            filename = photos.save(request.files['photo'])
+            data = request.form.to_dict()
+            data["image"] = filename
+            id_var = data_manager.add_question_or_answer(data)
+        except:
+            data = request.form.to_dict()
+            id_var = data_manager.add_question_or_answer(data)
+
+        return redirect('question/{}'.format(id_var))
 
 
 @app.route("/question/<int:question_id>", methods=['GET', 'POST'])
 def question(question_id):
     html_file = "question_with_answers.html"
     question, answers = data_manager.get_question_byid(question_id)
-    if request.method == 'POST' and 'photo' in request.files:
-        try:
-            filename = photos.save(request.files['photo'])
-            data = request.form.to_dict()
-            data["image"] = filename
-            data_manager.render_question_or_answer(data, question, question_id)
-            return redirect('question/{}'.format(question_id))
-        except:
-            data = request.form.to_dict()
-            data_manager.render_question_or_answer(data, question, question_id)
-            return redirect('question/{}'.format(question_id))
 
-    return render_template(html_file, question=question, answers=answers, q_head=q_head, a_head=a_head)
+    return render_template(html_file, question=question[0], answers=answers, q_head=q_head, a_head=a_head)
 
 
 @app.route('/question/<question_id>/new-answer', methods=['GET', 'POST'])
 def write_answer(question_id):
     submission_time = data_manager.time_generator()
     if request.method == 'POST':
-        return render_template('new_answer.html', next_id=0, question_id=question_id, submission_time=submission_time)
+        return render_template('new_answer.html', question_id=question_id, submission_time=submission_time)
 
 
 @app.route('/question/<question_id>/edit', methods=['GET', 'POST'])
@@ -94,19 +101,9 @@ def update_question(question_id):
 @app.route('/question/<int:question_id>/vote-<dir>', methods=["POST"])
 def voter(question_id, dir):
     operatorr = operator.__add__ if dir == "up" else operator.__sub__
-    question, answers = data_manager.get_question_byid(question_id)
     data = request.form.to_dict()
-    if "question" in data.keys():
-        data_manager.vote(question_id, question, data, operatorr)
-        return redirect('question/{}'.format(question_id))
-    elif "answer" in data.keys():
-        data_manager.vote(question_id, answers, data, operatorr)
-        return redirect('question/{}'.format(question_id))
-
-
-@app.template_filter('ctime')
-def timectime(s):
-    return time.ctime(int(s))
+    data_manager.vote(question_id, data, operatorr)
+    return redirect('question/{}'.format(question_id))
 
 
 if __name__ == '__main__':
